requirements.txt
Flask
joblib
PyPDF2
python-docx
scikit-learn
pytest
pytest-flask
pandas
